SimpleCov wasn't configured for Rails
Which meant that the coverage number included everything in `config/`
and `spec/`, all of which was run by the specs.

Also took some time to tweek the RSec config in general.

Addresses #88 (in that it reveals that coverage is actually not very
good)
