+++ .rspec
@@ -1,3 +1,2 @@
 --require spec_helper
 --color
---format documentation
