+++ spec/models/user_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe User do
+RSpec.describe User do
   before :each do
     @user = create :user, :admin
   end
